requirements.txt
Django==1.6.2
Jinja2==2.7.2
MarkupSafe==0.19
Pygments==1.6
Sphinx==1.2.2
argparse==1.2.1
bagit==1.3.5
docutils==0.11
python-magic==0.4.6
sphinx-rtd-theme==0.1.6
wsgiref==0.1.2
